scrollX type unified to CGFloat (her binding type); ContentView keeps
routing to SandboxView; her duplicate view-model stub dropped in favor
of the full coordinator.

+++ marinesandbox/Views/Canvas/SandboxView.swift
@@ -47,8 +47,8 @@ struct SandboxView: View {
                 let seabedY = geometry.size.height
                 ZStack(alignment: .bottomLeading) {
                     ParallaxScrollView(scrollX: Binding(
-                        get: { CGFloat(viewModel.scrollX) },
-                        set: { viewModel.scrollX = Double($0) }
+                        get: { viewModel.scrollX },
+                        set: { viewModel.scrollX = $0 }
                     ))
 
                     entityLayer(viewModel: viewModel, seabedY: seabedY)
